backend/services/watermark.py
"""
SportShield Watermark Service
Invisible DCT-based watermarking using imwatermark library.
Supports images and per-frame video watermarking.
"""
import os
import cv2
import numpy as np
from typing import Tuple, Dict, Optional
from PIL import Image

try:
    from imwatermark import WatermarkEncoder, WatermarkDecoder
    _WM_AVAILABLE = True
except ImportError:
    _WM_AVAILABLE = False
    print("[Watermark] imwatermark not available — using LSB fallback")


def _make_watermark_bits(watermark_id: str, asset_id: str) -> bytes:
    """Create a 32-byte watermark payload from IDs."""
    payload = f"{asset_id}:{watermark_id}"
    encoded = payload.encode("utf-8")[:32]
    return encoded.ljust(32, b"\x00")


def _lsb_embed(image: np.ndarray, payload: bytes) -> np.ndarray:
    """Simple LSB steganography fallback."""
    img = image.copy()
    bits = "".join(format(byte, "08b") for byte in payload)
    h, w, c = img.shape
    flat = img.flatten()
    for i, bit in enumerate(bits):
        flat[i] = (flat[i] & ~1) | int(bit)
    return flat.reshape(h, w, c)


def _lsb_decode(image: np.ndarray, length_bytes: int = 32) -> bytes:
    """Decode LSB-embedded payload."""
    flat = image.flatten()
    bits = [str(flat[i] & 1) for i in range(length_bytes * 8)]
    payload = bytes(int("".join(bits[i:i+8]), 2) for i in range(0, len(bits), 8))
    return payload


def embed_watermark(
    image: np.ndarray,
    watermark_id: str,
    asset_id: str,
    strength: float = 0.5,
) -> Tuple[np.ndarray, bool]:
    """
    Embed invisible watermark into an image.
    Returns (watermarked_image, success).
    """
    payload = _make_watermark_bits(watermark_id, asset_id)

    if _WM_AVAILABLE:
        try:
            # imwatermark expects BGR (OpenCV format)
            bgr = cv2.cvtColor(image, cv2.COLOR_RGB2BGR) if image.shape[2] == 3 else image
            encoder = WatermarkEncoder()
            encoder.set_watermark("bytes", payload)
            bgr_wm = encoder.encode(bgr, "dwtDct")
            rgb_wm = cv2.cvtColor(bgr_wm, cv2.COLOR_BGR2RGB)
            return rgb_wm, True
        except Exception as e:
            print(f"[Watermark] DCT embed failed: {e} — using LSB")

    # LSB fallback
    wm = _lsb_embed(image, payload)
    return wm, True


def detect_watermark(
    image: np.ndarray,
    expected_id: Optional[str] = None,
) -> Dict:
    """
    Detect watermark in an image.
    Returns {found: bool, payload: str, integrity: float}
    """
    if _WM_AVAILABLE:
        try:
            bgr = cv2.cvtColor(image, cv2.COLOR_RGB2BGR) if image.shape[2] == 3 else image
            decoder = WatermarkDecoder("bytes", 32 * 8)
            payload = decoder.decode(bgr, "dwtDct")
            payload_str = payload.decode("utf-8", errors="replace").rstrip("\x00")
            found = len(payload_str) > 3

            if expected_id and found:
                integrity = 1.0 if expected_id in payload_str else 0.5
            else:
                integrity = 0.9 if found else 0.0

            return {"found": found, "payload": payload_str, "integrity": integrity}
        except Exception as e:
            print(f"[Watermark] DCT detect failed: {e}")

    # LSB fallback
    try:
        payload = _lsb_decode(image)
        payload_str = payload.decode("utf-8", errors="replace").rstrip("\x00")
        found = len(payload_str) > 3
        integrity = 0.85 if found else 0.0
        return {"found": found, "payload": payload_str, "integrity": integrity}
    except Exception:
        return {"found": False, "payload": "", "integrity": 0.0}


def embed_watermark_to_video(
    input_path: str,
    output_path: str,
    watermark_id: str,
    asset_id: str,
    strength: float = 0.5,
) -> bool:
    """Embed watermark into every frame of a video."""
    cap = cv2.VideoCapture(input_path)
    if not cap.isOpened():
        return False

    fps = cap.get(cv2.CAP_PROP_FPS)
    width = int(cap.get(cv2.CAP_PROP_FRAME_WIDTH))
    height = int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT))

    fourcc = cv2.VideoWriter_fourcc(*"mp4v")
    out = cv2.VideoWriter(output_path, fourcc, fps, (width, height))

    frame_count = 0
    while True:
        ret, frame = cap.read()
        if not ret:
            break
        # Only watermark every 30th frame for speed; copy others
        if frame_count % 30 == 0:
            rgb = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
            wm_rgb, _ = embed_watermark(rgb, watermark_id, asset_id, strength)
            wm_bgr = cv2.cvtColor(wm_rgb, cv2.COLOR_RGB2BGR)
            out.write(wm_bgr)
        else:
            out.write(frame)
        frame_count += 1

    cap.release()
    out.release()
    return True


def resilience_profile(strength: int) -> Dict:
    """Return which attacks the watermark survives at given strength (1-10)."""
    checks = {
        "compression": strength >= 2,
        "resize": strength >= 3,
        "crop": strength >= 5,
        "color_grade": strength >= 7,
        "speed_change": strength >= 9,
    }
    return checks
