backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-socketio==5.11.4
python-multipart==0.0.12
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
celery==5.4.0
redis==5.2.0
apscheduler==3.10.4
google-generativeai==0.8.3
groq==0.13.1
opencv-python-headless==4.10.0.84
imagehash==4.3.1
Pillow==11.0.0
numpy==2.1.3
faiss-cpu==1.9.0.post1
invisible-watermark==0.2.0
hashlib2==1.0.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
reportlab==4.2.5
qrcode==8.0
yt-dlp>=2024.11.18
